+++ tests/files/src/0242_void_71.php
@@ -3,3 +3,4 @@ function f1() : void {}
 function f2() : void { return rand(0,2); }
 function f3() : int {}
 function f3() : ?string {}
+function f5() : ?string { return; }